build: altera a infraestutura do projeto para modularização
