+++ autou/settings.py
@@ -44,6 +44,8 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'api',
+    'emails',
+    'www'
 ]
 
 MIDDLEWARE = [
