+++ emails/templates/email_detail.html
@@ -1,4 +1,4 @@
-{% extends 'header.html' %}
+{% extends '../../api/templates/header.html' %}
 
 {% block content %}
     <body>
