--- media/Aqui está um exemplo de um email pr.txt
@@ -1,12 +1,0 @@
-Email: thecobsb@gmail.com
-Assunto: Status do projeto
-Mensagem:
-Prezado(a) [Nome do Destinatário],
-Espero que este email o(a) encontre bem. Gostaria de solicitar uma atualização sobre o status atual do projeto [Nome do Projeto]. Seria possível compartilhar os principais avanços realizados até o momento, bem como eventuais desafios ou pontos que necessitem de atenção?
-Além disso, gostaria de confirmar se há alguma ação pendente da minha parte ou da equipe para garantir que as próximas etapas ocorram conforme planejado. Caso precise de mais informações ou suporte adicional, por favor, não hesite em me informar.
-Agradeço desde já pela atenção e fico à disposição para quaisquer dúvidas ou esclarecimentos.
-Atenciosamente,
-[Seu Nome]
-[Seu Cargo]
-[Nome da Empresa]
-[Contato: telefone/email]
